Specify Model Attributes at runtime when creating tables.
Sometimes you are dealing with a Model library which cannot or should
not have dependencies on ServiceStack.Net, and/or you cannot or do not
want to decorate the Classes and Properties with data annotation
attributes. In the data access layer you may still want to make use of
the features the attributes provide in the Dal, such as the
AutoIncrmentAttribute, ForeignKeyAttribute, etc.

This commit adds support for this functionality via the new
DataAnnotationsCollection class, optionally accepted by CreateTable and
its overloads.

This feature furthers ServiceStack.OrmLite's goal of supporting true
POCO and not requiring attributes in the Business layer.