Empty field names in Join queries when not using Aliases in POCO properties
When the properties of POCO involved in joins don't have Aliases the generated query presents an empty string in the relative field name as in
SELECT * FROM "Customers" LEFT JOIN "Addresses" ON "Customers"."" = "Addresses"."". This change should fix the problem. If the Alias for a filed is not present it uses a field name the property name as in the name convention.