Implemented Sequences in SqlServer dialect provider.
Implemented Sequences in SqlServer dialect provider.

Fields with AutoIncrement attribute are now automatically included in the list of fields to return when gererating the insert sql statement.

Fields with AutoIncrement attribute and a Sequence attribute are also included in the insert field list with the value clause set to an sql expression to get the value from the sequence.
This reduces database calls by avoiding a separate call to get the auto-increment value since the insert call does it all now.

Fields with the Sequence attribute will no longer generate a IDENTITY field type. Instead it now generates a regular field with the default to "NEXT VALUE FOR {sequence_name}".
This also allows for safe inserts from other processes/apps.